fixing warning
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@6004 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ src/zgetrf_panel_nopiv_batched.cpp
@@ -29,7 +29,7 @@ magma_zgetrf_panel_nopiv_batched_q(
     //  panel factorization
     //===============================================
     if(m < nb){
-        printf("magma_zgetrf_panel_nopiv_batched_q m < nb %d < %d \n",m,nb);
+        printf("magma_zgetrf_panel_nopiv_batched_q m < nb %d < %d \n",(int) m, (int) nb);
         return -101;
     }
